pyproject.toml
[tool.poetry]
name = "evora-wheel"
version = "0.1.0"
description = "Controller and server for the Evora filter wheel"
authors = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
packages = [{include = "evora_wheel", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9,<4"
daemonocle = "^1.0.2"
click = "^8.1.3"
click-default-group = "^1.2.2"
phidget22 = "^1.14.20230331"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
ipython = ">=8.0.0"
ipdb = ">=0.13.3"
ruff = ">=0.4.0"

[tool.poetry.scripts]
evora-wheel = "evora_wheel.__main__:evora_wheel"

[tool.ruff]
line-length = 88
target-version = 'py312'
exclude = ["typings/"]

[ruff.lint]
select = ["E", "F", "I"]
unfixable = ["F841"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403", "E402"]

[tool.ruff.lint.isort]
known-first-party = ["evora_wheel"]
lines-after-imports = 2
section-order = ["future", "standard-library", "typing", "third-party", "sdss", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
typing = ["typing"]
sdss = ["sdsstools"]

[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
